rebase with handler changes
