+++ jupyterlab_bigquery/jupyterlab_bigquery/details_handler/__init__.py
@@ -1,1 +1,1 @@
-from .details_handler import DatasetDetailsHandler, TableDetailsHandler, TablePreviewHandler
+from .details_handler import DatasetDetailsHandler, TableDetailsHandler, TablePreviewHandler, format_preview_fields, format_preview_rows
