+++ jupyterlab_bigquery/jupyterlab_bigquery/details_handler/details_handler.py
@@ -6,7 +6,7 @@ import json
 import re
 import tornado.gen as gen
 import os
-import random
+import math
 
 import json
 import datetime
@@ -22,6 +22,7 @@ SCOPE = ("https://www.googleapis.com/auth/cloud-platform",)
 
 def create_bigquery_client():
   return bigquery.Client()
+
 
 def get_dataset_details(client, dataset_id):
   dataset = client.get_dataset(dataset_id)
@@ -54,6 +55,7 @@ def get_dataset_details(client, dataset_id):
       }
   }
 
+
 def format_detail_field(formatted_schema, field, field_name):
   if field.field_type == 'RECORD':
     formatted_schema.append({
@@ -61,51 +63,76 @@ def format_detail_field(formatted_schema, field, field_name):
         'type': 'RECORD',
         'description': field.description,
         'mode': field.mode
-      })
+    })
     for sub_field in field.fields:
-      format_detail_field(formatted_schema, sub_field, field_name + field.name + '.')
+      format_detail_field(formatted_schema, sub_field,
+                          field_name + field.name + '.')
   else:
     formatted_schema.append({
-      'name': field_name + field.name,
-      'type': field.field_type,
-      'description': field.description,
-      'mode': field.mode
+        'name': field_name + field.name,
+        'type': field.field_type,
+        'description': field.description,
+        'mode': field.mode
     })
+
 
 def format_detail_schema(schema):
   formatted_schema = []
   for field in schema:
     format_detail_field(formatted_schema, field, '')
   return formatted_schema
-    
+
 
 def get_table_details(client, table_id):
   table = client.get_table(table_id)
   return {
       'details': {
-          'id': "{}.{}.{}".format(table.project, table.dataset_id, table.table_id),
-          'name': table.table_id,
-          'description': table.description,
-          'labels': ["\t{}: {}".format(label, value) for label, value in table.labels.items()] if table.labels else None,
-          'date_created': json.dumps(table.created.strftime('%b %e, %G, %l:%M:%S %p'))[1:-1],
-          'expires': json.dumps(table.expires.strftime('%b %e, %G, %l:%M:%S %p'))[1:-1] if table.expires else None,
-          'location': table.location,
-          'last_modified': json.dumps(table.modified.strftime('%b %e, %G, %l:%M:%S %p'))[1:-1],
-          'project': table.project,
-          'dataset': table.dataset_id,
-          'link': table.self_link,
-          'num_rows': table.num_rows,
-          'num_bytes': table.num_bytes,
-          'schema': format_detail_schema(table.schema)
+          'id':
+              "{}.{}.{}".format(table.project, table.dataset_id,
+                                table.table_id),
+          'name':
+              table.table_id,
+          'description':
+              table.description,
+          'labels': [
+              "\t{}: {}".format(label, value)
+              for label, value in table.labels.items()
+          ] if table.labels else None,
+          'date_created':
+              json.dumps(table.created.strftime('%b %e, %G, %l:%M:%S %p'))
+              [1:-1],
+          'expires':
+              json.dumps(table.expires.strftime('%b %e, %G, %l:%M:%S %p'))[1:-1]
+              if table.expires else None,
+          'location':
+              table.location,
+          'last_modified':
+              json.dumps(table.modified.strftime('%b %e, %G, %l:%M:%S %p'))
+              [1:-1],
+          'project':
+              table.project,
+          'dataset':
+              table.dataset_id,
+          'link':
+              table.self_link,
+          'num_rows':
+              table.num_rows,
+          'num_bytes':
+              table.num_bytes,
+          'schema':
+              format_detail_schema(table.schema)
       }
   }
+
 
 def format_preview_field(formatted_fields, field, field_name):
   if field.field_type == 'RECORD':
     for record_entry in field.fields:
-      format_preview_field(formatted_fields, record_entry, field_name + field.name + ".")
+      format_preview_field(formatted_fields, record_entry,
+                           field_name + field.name + ".")
   else:
     formatted_fields.append(field_name + field.name)
+
 
 def format_preview_fields(schema):
   formatted_fields = []
@@ -113,39 +140,71 @@ def format_preview_fields(schema):
     format_preview_field(formatted_fields, field, '')
   return formatted_fields
 
-def format_preview_value(formatted_row, value):
-  if isinstance(value, bytes):
-      formatted_row.append(base64.b64encode(value).__str__()[2:-1])
-  elif isinstance(value, dict):
-    for sub_value in value.values():
-      format_preview_value(formatted_row, sub_value)
+
+def format_preview_value(value):
+  if value is None:
+    return None
+  elif isinstance(value, bytes):
+    return base64.b64encode(value).__str__()[2:-1]
   elif isinstance(value, float):
     if value == float('inf'):
-      formatted_row.append('Infinity')
+      return 'Infinity'
     elif value == float('-inf'):
-      formatted_row.append('-Infinity')
+      return '-Infinity'
     elif math.isnan(value):
-      formatted_row.append('NaN')
+      return 'NaN'
     else:
-      formatted_row.append(value.__str__())
+      return value
   elif isinstance(value, datetime.datetime):
-    formatted_row.append(json.dumps(value.strftime('%Y-%m-%d %H:%M:%S.%f %Z'))[1:-1])
+    return json.dumps(value.strftime('%Y-%m-%d %H:%M:%S.%f %Z'))[1:-1]
   else:
-    formatted_row.append(value.__str__())
+    return value.__str__()
 
-def format_preview_row(row):
-  formatted_row = []
-  for value in row.values():
-    format_preview_value(formatted_row, value)
-  return formatted_row
-  
+
 def get_table_preview(client, table_id):
   table = client.get_table(table_id)
   rows = client.list_rows(table, max_results=100)
+  fields = rows.schema
+
   return {
-    'fields': format_preview_fields(rows.schema),
-    'rows': [format_preview_row(row) for row in rows]
+      'fields': format_preview_fields(fields),
+      'rows': format_preview_rows(rows, fields)
   }
+
+
+def format_preview_rows(rows, fields):
+  formatted_rows = []
+  for row in rows:
+    formatted_row = []
+    for i in range(len(fields)):
+      check_repeated(row[i], fields[i], formatted_row)
+    formatted_rows.append(formatted_row)
+  return formatted_rows
+
+
+def check_repeated(value, field, formatted_row):
+  if field.mode == 'REPEATED':
+    if len(value) > 0:
+      handle_records(value[0], field, formatted_row)
+    else:
+      handle_records(None, field, formatted_row)
+  else:
+    handle_records(value, field, formatted_row)
+
+
+def handle_records(value, field, formatted_row):
+  if field.field_type == 'RECORD':
+    if value is None:
+      for sub_field in field.fields:
+        handle_records(None, sub_field, formatted_row)
+    else:
+      for sub_field in field.fields:
+        if value[sub_field.name]:
+          check_repeated(value[sub_field.name], sub_field, formatted_row)
+        else:
+          handle_records(None, sub_field, formatted_row)
+  else:
+    formatted_row.append(format_preview_value(value))
 
 
 class DatasetDetailsHandler(APIHandler):
@@ -181,11 +240,8 @@ class TableDetailsHandler(APIHandler):
     except Exception as e:
       app_log.exception(str(e))
       self.set_status(500, str(e))
-      self.finish({
-          'error': {
-              'message': str(e)
-          }
-      })
+      self.finish({'error': {'message': str(e)}})
+
 
 class TablePreviewHandler(APIHandler):
   """"Handles request for table preview."""
@@ -202,8 +258,4 @@ class TablePreviewHandler(APIHandler):
     except Exception as e:
       app_log.exception(str(e))
       self.set_status(500, str(e))
-      self.finish({
-          'error': {
-              'message': str(e)
-          }
-      })
+      self.finish({'error': {'message': str(e)}})
