+++ jupyterlab_bigquery/jupyterlab_bigquery/pagedAPI_handler/pagedQuery_handler.py
@@ -1,8 +1,7 @@
 from jupyterlab_bigquery.pagedAPI_handler import PagedAPIHandler
 from google.cloud import bigquery
 import json
-from jupyterlab_bigquery.handlers import format_preview_fields, format_preview_rows
-
+from jupyterlab_bigquery.details_handler import format_preview_fields, format_preview_rows
 
 
 class PagedQueryHandler(PagedAPIHandler):
